Use array method include and filter for path and routes for rendering specific data using id. Use split to get the id inside renderDetails component. Modified css.

+++ js/components/Details/Details.js
@@ -1,15 +1,17 @@
 export const Details = {
     render: () => {
         return `
-            <div class="details">
+            <div class="details post">
             </div>
-            <button class="delete">delete</button>
+            <div class="btn">
+                <button class="delete">DELETE</button>
+            </div>
         `;
     }
 }
 
 export const renderDetails = async () => {
-    const id = new URLSearchParams(window.location.search).get('id');
+    const id = window.location.hash.split('=')[1];
     const container = document.querySelector('.details');
     const deleteBtn = document.querySelector('.delete');
 
@@ -22,13 +24,14 @@ export const renderDetails = async () => {
     `
 
     container.innerHTML = template;
-
+    
     deleteBtn.addEventListener('click', async () => {
-        const res = await fetch('http://localhost:8888/posts/' + id, {
-            method: 'DELETE'
-        });
-
-        window.location.href = '/';
+        if(confirm("Are you sure to want to delete this blog?")) {
+            await fetch('http://localhost:8888/posts/' + id, {
+                method: 'DELETE'
+            });
+            window.location.href = '/';
+        }
     })
 
     window.addEventListener('DOMContentLoaded', () => renderDetails());
